Cleaned up history ring buffer code, including array out-of-bounds fix
When 'historypresent' is 0, the previous element is at the end of the
buffer; the old code would look up index -1 which caused a crash on my
x86_64 Linux system.

Added a bit of documentation about the ring buffer.

Use 'sizeof(glui32)' instead of hardcoded '4'.

No need to check for NULL before calling 'free', since 'free(NULL)' is
allowed.

Old entries were being freed both before and after inserting a new
element; I removed the latter.

Indices must always be in the range [0..HISTORYLEN) and are increased
in increments of one, so they will always wrap to index 0; made that
explicit in the code.